finished run on model evaluation with attached liked data, setting things up so that user data can be gathered with likes and synthetic ratings can be attached, moving more of those methods into get_user_ratings
