+++ data_processing/get_user_ratings.py
@@ -4,6 +4,7 @@ import asyncio
 import datetime
 import os
 from pprint import pprint
+from statistics import mean
 
 import requests
 from bs4 import BeautifulSoup
@@ -54,7 +55,67 @@ def get_page_count(username):
     return num_pages, display_name
 
 
-def get_user_data(username, data_opt_in=False):
+def attach_synthetic_ratings(
+    all_ratings: list, global_mean: float, global_weight: int = 5
+) -> None:
+    """
+    Mutates `all_ratings` in place, adding `synthetic_rating_val`:
+        - If rating_val >= 0: synthetic = rating_val
+        - Else if liked == True: synthetic = weighted mean of (user liked+rated mean, global liked+rated mean)
+        - Else: leave unset
+    """
+
+    # 1) Per-user mean over liked & rated
+    liked_rated_by_user = {}
+    for r in all_ratings:
+        uid = r.get("user_id")
+        if not uid:
+            continue
+        if (
+            r.get("liked") is True
+            and isinstance(r.get("rating_val"), (int, float))
+            and r["rating_val"] >= 0
+        ):
+            liked_rated_by_user.setdefault(uid, []).append(r["rating_val"])
+
+    user_mean_cache = {
+        uid: (mean(vals), len(vals)) for uid, vals in liked_rated_by_user.items()
+    }
+
+    # 2) Attach synthetic per row
+    for r in all_ratings:
+        rv = r.get("rating_val")
+        if isinstance(rv, (int, float)) and rv >= 0:
+            # Explicit rating: prefer it
+            r["synthetic_rating_val"] = rv
+            continue
+
+        # Only compute synthetic for unrated-but-liked
+        if r.get("liked") is not True:
+            continue
+
+        uid = r.get("user_id")
+        u_mean, u_n = user_mean_cache.get(uid, (None, 0))
+
+        # Decide synthetic value
+        synthetic = None
+        if u_mean is not None and global_mean is not None:
+            synthetic = (u_n * u_mean + global_weight * global_mean) / (
+                u_n + global_weight
+            )
+        elif u_mean is not None:
+            synthetic = u_mean
+        elif global_mean is not None:
+            synthetic = global_mean
+        else:
+            synthetic = None
+
+        if synthetic is not None:
+            # keep as float; if you prefer ints, you could round here
+            r["synthetic_rating_val"] = float(synthetic)
+
+
+def get_user_data(username, data_opt_in=False, include_liked_items=True):
     num_pages, display_name = get_page_count(username)
     if num_pages == -1:
         return [], "user_not_found"
@@ -68,21 +129,29 @@ def get_user_data(username, data_opt_in=False):
             store_in_db=False,
             num_pages=num_pages,
             return_unrated=True,
-            attach_liked_flag=True,
+            attach_liked_flag=include_liked_items,
         )
     )
     loop.run_until_complete(future)
 
-    user_ratings = [x for x in future.result() if x["rating_val"] >= 0]
-
+    user_ratings = [x for x in future.result()]
+    explicit_user_ratings = [x for x in user_ratings if x["rating_val"] >= 0]
     if data_opt_in:
         # Remove "liked" flag, if it exists, before sending to DB
         sanitized_ratings = [
-            {k: v for k, v in r.items() if k != "liked"} for r in user_ratings
+            {k: v for k, v in r.items() if k != "liked"} for r in explicit_user_ratings
         ]
         send_to_db(username, display_name, user_ratings=sanitized_ratings)
 
-    return future.result(), "success"
+    if include_liked_items:
+        user_ratings_with_likes = [
+            x for x in user_ratings if x["rating_val"] >= 0 or x["liked"]
+        ]
+        attach_synthetic_ratings(user_ratings_with_likes, global_mean=7.9826)
+
+    print(user_ratings_with_likes)
+
+    return user_ratings_with_likes, "success"
 
 
 def send_to_db(username, display_name, user_ratings):
