+++ data_processing/models/eval_results/best_svd_params.json
@@ -1,10 +1,10 @@
 {
-  "init_std_dev": 0.21483583767866765,
-  "lr_all": 0.004193789482141866,
+  "init_std_dev": 0.22893522915121478,
+  "lr_all": 0.009526972340976523,
   "n_epochs": 69,
-  "n_factors": 217,
-  "reg_bi": 0.12064297187296281,
-  "reg_bu": 0.04830398022450677,
-  "reg_pu": 0.0460519442989215,
-  "reg_qi": 0.04858665178520828
+  "n_factors": 150,
+  "reg_bi": 0.09342858816671092,
+  "reg_bu": 0.06765045263862099,
+  "reg_pu": 0.047550025826647725,
+  "reg_qi": 0.04790153247439678
 }
